Normalize `repo_utils` subpackage
Same import pattern as others. Also hook into docs.
